Removing the best_gpu() function and renaming to demo.rs
